Have subnet outputs depend on NAT routes
This ensures that if NAT gateways are requested, that the module does
not give the subnets as outputs until the NAT routes are added, and
hence the network is actually routeable. This is important if you are
bootstrapping a cluster (ie: an ASG) over the internet to prevent race
conditions.