private_subnets: Corrected list notation to element function
Fixed the private_subnets resource so that index does not go out of
range when computing a number of subnets larger than the available
availability zones.  This was due to this resource using list notation
instead of the element interpolation function. List notation will go
out of range, but by design element will wrap.